+++ src/commands/index.ts
@@ -5,6 +5,7 @@ import { registerEditorCommands } from "./editor";
 import { registerFileCommands } from "./file";
 import { registerFollowCommands } from "./follow";
 import { registerGistCommands } from "./gist";
+import { registerGistLogCommands } from "./gistLog";
 import { registerPasteImageCommands } from "./pasteImage";
 
 export function registerCommands(context: ExtensionContext) {
@@ -13,6 +14,7 @@ export function registerCommands(context: ExtensionContext) {
   registerEditorCommands(context);
   registerFollowCommands(context);
   registerGistCommands(context);
+  registerGistLogCommands(context);
   registerFileCommands(context);
   registerPasteImageCommands(context);
 }